Split FlowsSection list vs detail (552 → 175 lines)
- FlowsSection.tsx        list view shell — refresh, create, filter,
                          navigates to detail
- flows/FlowRow.tsx       list row tile
- flows/FlowDetail.tsx    editor — metadata fields + stage list + save
                          and lifecycle actions (publish / archive /
                          duplicate)
- flows/StageRow.tsx      collapsible per-stage editor
- flows/shared.tsx        STATUS_STYLE constant, View type, Field
                          atom (used by FlowDetail + StageRow)

All files under 240 lines.
